refactor: extract logic of adding player to helper
For the new endpoint to add bots to a game the
same logic is going to be required
